src/google-cloud/projects/no-sql-writer/app/requirements.txt
flask
google-cloud-firestore
python-dotenv
gunicorn
